Implement routability more or less completely

+++ tests/common/location.rs
@@ -3,6 +3,6 @@ use serde::Deserialize;
 #[derive(Clone, Copy, Debug, Default, Deserialize)]
 pub struct Location {
     // Note: The order is important since we derive Deserialize
-    pub longitude: f32,
-    pub latitude: f32,
+    pub longitude: f64,
+    pub latitude: f64,
 }
